Carpooling service created as a dedicated service

+++ y-a-foot-api/src/main/java/net/andresbustamante/yafoot/services/SiteManagementService.java
@@ -17,5 +17,5 @@ public interface SiteManagementService {
      * @return New site's identifier
      * @throws DatabaseException
      */
-    long saveSite(Site site, UserContext userContext) throws DatabaseException;
+    Integer saveSite(Site site, UserContext userContext) throws DatabaseException;
 }
